* #2370 clean compile warning
* update changelog

+++ core/src/db/IDGenerator.cpp
@@ -47,7 +47,7 @@ SimpleIDGenerator::NextIDNumbers(size_t n, IDNumbers& ids) {
     auto micros = std::chrono::duration_cast<std::chrono::microseconds>(now.time_since_epoch()).count();
     micros *= MAX_IDS_PER_MICRO;
 
-    for (int pos = 0; pos < n; ++pos) {
+    for (size_t pos = 0; pos < n; ++pos) {
         ids.push_back(micros + pos);
     }
     return Status::OK();
@@ -114,7 +114,7 @@ SafeIDGenerator::NextIDNumbers(size_t n, IDNumbers& ids) {
 
     int64_t ID_high_part = time_stamp_ms_ * MAX_IDS_PER_MICRO;
 
-    for (int pos = 0; pos < n; ++pos) {
+    for (size_t pos = 0; pos < n; ++pos) {
         ids.push_back(ID_high_part + pos);
     }
